refactor(cmd): rename cmd upgrade to cmd switch
