+++ src/cmd/switch.rs
